Move interface to other locations
- adapters to gridtools/storage/adapter
- layout_transformation to gridtools/layout_transformation
