+++ tests/unit_tests/CMakeLists.txt
@@ -41,4 +41,3 @@ add_subdirectory(distributed_boundaries)
 add_subdirectory(stencil_composition)
 add_subdirectory(storage)
 add_subdirectory(layout_transformation)
-add_subdirectory(interface)
